+++ src/core/injections/results_based/techniques/classic/cb_handler.py
@@ -286,9 +286,12 @@ def cb_injection_handler(url, timesec, filename, http_request_method):
               vp_flag = logs.add_parameter(vp_flag, filename, the_type, header_name, http_request_method, vuln_parameter, payload)
             logs.update_payload(filename, counter, payload) 
             counter = counter + 1
-            
-            if not settings.VERBOSITY_LEVEL >= 1 and not settings.LOAD_SESSION:
-              print ""
+
+            if not settings.LOAD_SESSION:
+              if not settings.VERBOSITY_LEVEL >= 1:
+                print ""
+              else:
+                checks.total_of_requests()
 
             # Print the findings to terminal.
             success_msg = "The"