+++ src/core/injections/semiblind/techniques/file_based/fb_handler.py
@@ -515,8 +515,11 @@ def fb_injection_handler(url, timesec, filename, http_request_method, url_time_r
             logs.update_payload(filename, counter, payload) 
             counter = counter + 1
 
-            if not settings.VERBOSITY_LEVEL >= 1 and not settings.LOAD_SESSION:
-              print ""
+            if not settings.LOAD_SESSION:
+              if not settings.VERBOSITY_LEVEL >= 1:
+                print ""
+              else:
+                checks.total_of_requests()
 
             # Print the findings to terminal.
             success_msg = "The"