+++ src/core/injections/semiblind/techniques/tempfile_based/tfb_handler.py
@@ -448,8 +448,11 @@ def tfb_injection_handler(url, timesec, filename, tmp_path, http_request_method,
               logs.update_payload(filename, counter, payload) 
               counter = counter + 1
 
-              if not settings.VERBOSITY_LEVEL > 1 and not settings.LOAD_SESSION:
-                print ""
+              if not settings.LOAD_SESSION:
+                if not settings.VERBOSITY_LEVEL >= 1:
+                  print ""
+                else:
+                  checks.total_of_requests()
 
               # Print the findings to terminal.
               success_msg = "The"